Implement helper methods for CLI
At this point I believe I will have to shift these to another file in order to test them properly, instead of nesting them.
For the most part, this commit is just to save my progress.

All that's left to do should be to implement the options for terminal use